Passes the frame time as the update delta time
The game loop is fixed time step therefore it doesn't make sense to pass anything but the frame time

+++ block_builder/shared/src/commonMain/kotlin/com/edwardharker/kge/blockbuilder/system/AddBlockFailSystem.kt
@@ -17,7 +17,7 @@ import com.edwardharker.kge.util.Colour.Companion.WHITE
 import com.edwardharker.kge.util.Vector2
 
 object AddBlockFailSystem : UpdateSystem {
-    override fun update(world: World, deltaTime: Long) {
+    override fun update(world: World, deltaTime: Float) {
         world.forEachEntityWithComponents { entity: Entity,
                                             pointer: PointerComponent,
                                             collisionComponent: CollisionComponent ->
