+++ block_builder/shared/src/commonMain/kotlin/com/edwardharker/kge/blockbuilder/system/GameSuccessSystem.kt
@@ -16,7 +16,7 @@ import com.edwardharker.kge.util.Colour.Companion.WHITE
 import com.edwardharker.kge.util.Vector2
 
 object GameSuccessSystem : UpdateSystem {
-    override fun update(world: World, deltaTime: Long) {
+    override fun update(world: World, deltaTime: Float) {
         world.forEachEntityWithComponents { entity,
                                             transform: TransformComponent,
                                             rect: RectangleSpriteComponent,
