+++ engine/src/commonMain/kotlin/com/edwardharker/kge/system/UpdateSystem.kt
@@ -3,5 +3,5 @@ package com.edwardharker.kge.system
 import com.edwardharker.kge.World
 
 interface UpdateSystem {
-    fun update(world: World, deltaTime: Long)
+    fun update(world: World, deltaTime: Float)
 }
